# openapi.yaml
openapi: 3.0.3
info:
  title: Fly-Fleet API
  description: Private Jet Charter Brokerage Platform API
  version: 1.0.0
  contact:
    name: Fly-Fleet Support
    email: contact@fly-fleet.com
  license:
    name: Private
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/webhooks/resend:
    get:
      summary: List webhooks
      description: Handle external service webhooks
      tags:
        - webhooks
      responses:
        '200':
          description: Success
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Create webhooks
      description: Handle external service webhooks
      tags:
        - webhooks
      responses:
        '200':
          description: Success
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/whatsapp/link:
    get:
      summary: List whatsapp
      description: Generate WhatsApp communication links
      tags:
        - whatsapp
      responses:
        '200':
          description: Success
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Create whatsapp
      description: Generate WhatsApp communication links
      tags:
        - whatsapp
      responses:
        '200':
          description: Success
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin authentication using Bearer token with x-admin-email header
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Validation failed
              details:
                type: array
                items:
                  type: object
                  properties:
                    code:
                      type: string
                    path:
                      type: array
                      items:
                        oneOf:
                          - type: string
                          - type: integer
                    message:
                      type: string
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Rate limit exceeded. Maximum 7 requests per hour.
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Internal server error